requirements.txt
python-telegram-bot==21.1.1
requests==2.31.0
yt-dlp==2024.4.9